requirements.txt
Flask
fire
tqdm
pandas
requests
sqlalchemy 
psycopg2-binary
pyeuropeana
ibm-cos-sdk